Deb || alpha dtype change

+++ src/kernels/AttentionForwardKernelBatch.py
@@ -37,14 +37,13 @@ def _attention_forward_inner_mask(acc, l_i, m_i, q, desc_k, desc_v,
         # -- compute correction factor
         alpha = tl.math.exp2(m_i - m_ij)
         l_ij = tl.sum(p, 1)
-
+        alpha = alpha.to(dtype)
         acc = acc * alpha[:, None]
 
         # print(f"Offset of v [0, {offsetv_y}]")
         v = desc_v.load([offsetv_y, 0])
         p = p.to(dtype)
         acc = tl.dot(p, v, acc)
-        acc = acc.to(dtype)
         l_i = l_i * alpha + l_ij
         m_i = m_ij
         offsetk_y += block_n
